+++ src/functionalTest/groovy/com/autonomousapps/jvm/projects/AbiExclusionsProject.groovy
@@ -65,7 +65,7 @@ class AbiExclusionsProject extends AbstractProject {
   ]
 
   List<Advice> actualAdvice() {
-    return AdviceHelper.actualAdvice(gradleProject)
+    return AdviceHelper.actualAdviceForFirstSubproject(gradleProject)
   }
 
   final List<Advice> expectedAdvice = []