+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Dependency Analysis Plugin Changelog
 # Version 0.44.0 (unreleased)
 * [New] Can now filter/configure compileOnly advice, analogous to other kinds of advice.
 (Thank you to Stephane Nicolas @stephanenicolas for contributing this!)
-* [New] Improved support for Android variants in dependency analysis.
+* [New] Improved support for Android variants in dependency analysis. Plugin now analyzes all 
+source sets / variants for each execution of `buildHealth`.
 * [Fixed] Build cache issue with `ArtifactsAnalysisTask`. (Thanks to @Paladinium for filing the 
 issue and providing a reproducer!)